fix: rename interactive components to kebab-case and update imports to resolve IDE module resolution errors
