+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import { projects } from "@/data/projects"
 import { ProjectGrid } from "@/components/ProjectGrid"
 import { getRepoStats } from "@/lib/github"
-import { HomeWrapper } from "@/components/sections/HomeWrapper"
+import { HomeWrapper } from "../components/sections/home-wrapper"
 
 export default async function Home() {
   const stats = await getRepoStats()
