+++ src/components/sections/hero-content.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import { motion } from "framer-motion"
 import { ChevronDown } from "lucide-react"
-import { RepoStatsDisplay } from "./RepoStatsDisplay"
+import { RepoStatsDisplay } from "./repo-stats-display"
 import { GitHubStats } from "@/lib/github"
 
 interface HeroContentProps {
